refactor(layout): finalize facade and split layout store tests (#1080)
Move loadLayout/createNewLayout into layout/layout-lifecycle.ts, the
hasStarted/BackupState persistence helpers into layout/persistence.ts,
and the device-type query helpers into layout/queries.ts, leaving
layout.svelte.ts as a thin orchestrator. Extend LayoutStateAccess with
resetBackupTracking() so the lifecycle module can reset export tracking
through the existing state seam.

Split the 2718-line layout-store.test.ts into focused suites by concern
(rack-actions, device-type-actions, device-actions), keeping lifecycle
and misc cases in layout-store.test.ts. Pure move of existing tests;
case count unchanged at 143.
